Avances Profesor Inicio

+++ src/main/java/com/proyecto/spring/aulanote/aulanote/service/TareasService.java
@@ -26,23 +26,36 @@ public class TareasService {
     public Optional<Tareas> obtenerTareaPorId(Integer id) {
         return tareasRepository.findById(id);
     }
+   
 
-    public Optional<Tareas> actualizarTarea(Integer id, Tareas tareaDetalles) {
-        return tareasRepository.findById(id).map(tarea -> {
-            tarea.setNombreActividad(tareaDetalles.getNombreActividad());
-            tarea.setDescripcion(tareaDetalles.getDescripcion());
-            tarea.setFechaLimite(tareaDetalles.getFechaLimite());
-            tarea.setProfesorId(tareaDetalles.getProfesorId());
-            return tareasRepository.save(tarea);
-        });
-    }
 
-    public Boolean eliminarTarea(Integer id) {
-        return tareasRepository.findById(id).map(tarea -> {
-            tareasRepository.delete(tarea);
-            return true;
-        }).orElse(false);
-    }
+    public List<Tareas> listarPorProfesor(Integer profesorId) {
+        return tareasRepository.findByProfesorId(profesorId);
+}
+
+public Optional<Tareas> actualizarTarea(Integer id, Tareas tareaDetalles, Integer profesorId) {
+    return tareasRepository.findById(id).map(tarea -> {
+        if (!tarea.getProfesorId().equals(profesorId)) {
+            throw new RuntimeException("❌ No puedes modificar esta tarea, no es tuya.");
+        }
+        tarea.setNombreActividad(tareaDetalles.getNombreActividad());
+        tarea.setDescripcion(tareaDetalles.getDescripcion());
+        tarea.setFechaLimite(tareaDetalles.getFechaLimite());
+        return tareasRepository.save(tarea);
+    });
+}
+
+public Boolean eliminarTarea(Integer id, Integer profesorId) {
+    return tareasRepository.findById(id).map(tarea -> {
+        if (!tarea.getProfesorId().equals(profesorId)) {
+            throw new RuntimeException("❌ No puedes eliminar esta tarea, no es tuya.");
+        }
+        tareasRepository.delete(tarea);
+        return true;
+    }).orElse(false);
+}
+
+
 
     // --- Métodos personalizados ---
     public Optional<Tareas> calificarTarea(Integer id, String calificacion) {
